Moved logging back to the utils directory. Travis error looks bogus
