--- infoset/log/__init__.py
@@ -1,1 +1,0 @@
-"""Infoset log package."""
